Split MigrationAid.
